moved test classes to separate services
