+++ application/src/test/java/com/musicforall/services/songlist/SonglistServiceTest.java
@@ -1,4 +1,4 @@
-package com.musicforall;
+package com.musicforall.services.songlist;
 
 import com.musicforall.model.Song;
 import com.musicforall.model.Songlist;
